requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.40.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
